requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyserial>=3.5
pytest>=8.0.0
